+++ packages/dom-snapshot/src/browser/processPage.js
@@ -5,7 +5,7 @@ const aggregateResourceUrlsAndBlobs = require('./aggregateResourceUrlsAndBlobs')
 const makeGetResourceUrlsAndBlobs = require('./getResourceUrlsAndBlobs');
 const makeProcessResource = require('./processResource');
 const makeExtractResourcesFromSvg = require('./makeExtractResourcesFromSvg');
-const fetchUrl = require('./fetchUrl');
+const makeFetchUrl = require('./fetchUrl');
 const makeFindStyleSheetByUrl = require('./findStyleSheetByUrl');
 const makeExtractResourcesFromStyleSheet = require('./extractResourcesFromStyleSheet');
 const extractResourceUrlsFromStyleAttrs = require('./extractResourceUrlsFromStyleAttrs');
@@ -24,7 +24,10 @@ const makeLog = require('./log');
 const noop = require('./noop');
 const makeSessionCache = require('./sessionCache');
 
-function processPage(doc = document, {showLogs, useSessionCache, dontFetchResources} = {}) {
+function processPage(
+  doc = document,
+  {showLogs, useSessionCache, dontFetchResources, fetchTimeout} = {},
+) {
   /* MARKER FOR TEST - DO NOT DELETE */
   const log = showLogs ? makeLog(Date.now()) : noop;
   log('processPage start');
@@ -37,6 +40,7 @@ function processPage(doc = document, {showLogs, useSessionCache, dontFetchResour
   );
 
   const extractResourcesFromSvg = makeExtractResourcesFromSvg({extractResourceUrlsFromStyleTags});
+  const fetchUrl = makeFetchUrl({timeout: fetchTimeout});
   const processResource = makeProcessResource({
     fetchUrl,
     findStyleSheetByUrl,