+++ packages/dom-snapshot/src/browser/processResource.js
@@ -27,6 +27,9 @@ function makeProcessResource({
         const resourceUrls = getDependencies(url);
         log('doProcessResource from sessionStorage', url, 'deps:', resourceUrls.slice(1));
         cache[url] = Promise.resolve({resourceUrls});
+      } else if (/https:\/\/fonts.googleapis.com/.test(url)) {
+        log('not processing google font:', url);
+        cache[url] = Promise.resolve({resourceUrls: [url]});
       } else {
         const now = Date.now();
         cache[url] = doProcessResource(url).then(result => {
@@ -44,15 +47,26 @@ function makeProcessResource({
         .catch(e => {
           if (probablyCORS(e)) {
             return {probablyCORS: true, url};
+          } else if (e.isTimeout) {
+            return {isTimeout: true, url};
           } else {
             throw e;
           }
         })
-        .then(({url, type, value, probablyCORS}) => {
+        .then(({url, type, value, probablyCORS, isTimeout}) => {
           if (probablyCORS) {
             log('not fetched due to CORS', `[${Date.now() - now}ms]`, url);
             sessionCache && sessionCache.setItem(url, []);
             return {resourceUrls: [url]};
+          }
+
+          if (isTimeout) {
+            // TODO return errorStatusCode once VG supports it (https://trello.com/c/J5lBWutP/92-when-capturing-dom-add-non-200-urls-to-resource-map)
+            log('not fetched due to timeout, returning empty resource');
+            sessionCache && sessionCache.setItem(url, []);
+            return {
+              blobsObj: {[url]: {type: 'application/x-applitools-empty', value: new ArrayBuffer()}},
+            };
           }
 
           log(`fetched [${Date.now() - now}ms] ${url} bytes: ${value.byteLength}`);