+++ packages/sdk-shared/src/process-commons.js
@@ -1,10 +1,10 @@
-'use strict';
-const {promisify: p} = require('util');
-const path = require('path');
-const fs = require('fs');
-const os = require('os');
-const {spawn} = require('child_process');
-const {makeError} = require('@applitools/functional-commons');
+'use strict'
+const {promisify: p} = require('util')
+const path = require('path')
+const fs = require('fs')
+const os = require('os')
+const {spawn} = require('child_process')
+const {makeError} = require('@applitools/functional-commons')
 
 async function callFunctionViaProcess(
   jsFile,
@@ -15,11 +15,11 @@ async function callFunctionViaProcess(
   const jsModuleThatCallsExportedFunction = path.join(
     __dirname,
     'resources/js-module-that-calls-exported-function.js',
-  );
+  )
 
-  const tmpDir = await p(fs.mkdtemp)(os.tmpdir() + '/');
-  const resultFile = path.join(tmpDir, 'result.json');
-  const resultOutputFileDescriptor = await p(fs.open)(resultFile, 'w');
+  const tmpDir = await p(fs.mkdtemp)(os.tmpdir() + '/')
+  const resultFile = path.join(tmpDir, 'result.json')
+  const resultOutputFileDescriptor = await p(fs.open)(resultFile, 'w')
 
   try {
     await executeJsModule(
@@ -32,40 +32,40 @@ async function callFunctionViaProcess(
           stdio: ['inherit', 'inherit', 'inherit', resultOutputFileDescriptor],
         },
       },
-    );
+    )
 
-    await p(fs.close)(resultOutputFileDescriptor);
+    await p(fs.close)(resultOutputFileDescriptor)
 
-    return JSON.parse(await p(fs.readFile)(resultFile));
+    return JSON.parse(await p(fs.readFile)(resultFile))
   } catch (err) {
-    await p(fs.close)(resultOutputFileDescriptor);
+    await p(fs.close)(resultOutputFileDescriptor)
 
-    const errorString = await p(fs.readFile)(resultFile, 'utf-8');
+    const errorString = await p(fs.readFile)(resultFile, 'utf-8')
 
     if (errorString)
       throw makeError(new Error(errorString), {
         stdout: err.stdout,
         stderr: err.stderr,
         exitCode: err.exitCode,
-      });
+      })
     else {
-      throw err;
+      throw err
     }
   } finally {
     try {
-      await p(fs.unlink)(resultFile);
-      await p(fs.rmdir)(tmpDir);
+      await p(fs.unlink)(resultFile)
+      await p(fs.rmdir)(tmpDir)
     } catch (err) {
       console.error(
         'Oops. Failed to delete temporary file. Not too bad, so ignoring. File:',
         resultFile,
-      );
+      )
     }
   }
 }
 
 async function executeJsModule(jsModule, args = [], {spawnOptions = {}, timeout = undefined} = {}) {
-  return executeProcess(process.execPath, [jsModule].concat(args || []), {spawnOptions, timeout});
+  return executeProcess(process.execPath, [jsModule].concat(args || []), {spawnOptions, timeout})
 }
 
 function executeAndControlProcess(
@@ -76,7 +76,7 @@ function executeAndControlProcess(
   const subProcess = spawn(processPath, args, {
     stdio: 'pipe',
     ...spawnOptions,
-  });
+  })
 
   const exitPromise = new Promise((resolve, reject) => {
     subProcess.on('error', reject).on('close', (exitCode, signal) =>
@@ -111,20 +111,20 @@ function executeAndControlProcess(
               },
             ),
           ),
-    );
+    )
 
-    let stdout = subProcess.stdout ? '' : undefined;
-    let stderr = subProcess.stderr ? '' : undefined;
-    subProcess.stdout && subProcess.stdout.on('data', data => (stdout += data.toString()));
-    subProcess.stderr && subProcess.stderr.on('data', data => (stderr += data.toString()));
+    let stdout = subProcess.stdout ? '' : undefined
+    let stderr = subProcess.stderr ? '' : undefined
+    subProcess.stdout && subProcess.stdout.on('data', data => (stdout += data.toString()))
+    subProcess.stderr && subProcess.stderr.on('data', data => (stderr += data.toString()))
 
     if (timeout) {
-      setTimeout(() => subProcess.kill(), timeout);
+      setTimeout(() => subProcess.kill(), timeout)
     }
-    return {stdout, stderr};
-  });
+    return {stdout, stderr}
+  })
 
-  return {subProcess, exitPromise};
+  return {subProcess, exitPromise}
 }
 
 async function executeProcess(
@@ -132,29 +132,29 @@ async function executeProcess(
   args = [],
   {spawnOptions = {}, timeout = undefined} = {},
 ) {
-  return await executeAndControlProcess(processPath, args, {spawnOptions, timeout}).exitPromise;
+  return await executeAndControlProcess(processPath, args, {spawnOptions, timeout}).exitPromise
 }
 
 async function sh(command, options) {
   return await executeProcess('/bin/bash', ['-c', command], {
     ...options,
     spawnOptions: {stdio: 'inherit', ...(options || {}).spawnOptions},
-  });
+  })
 }
 
 async function shWithOutput(command, options, {withStderr = false} = {}) {
   const {stdout, stderr} = await sh(command, {
     ...options,
     spawnOptions: {stdio: 'pipe', ...(options || {}).spawnOptions},
-  });
+  })
 
-  return toArray(stdout).concat(withStderr ? toArray(stderr) : []);
+  return toArray(stdout).concat(withStderr ? toArray(stderr) : [])
 
   function toArray(std) {
     return std
       .trim()
       .split('\n')
-      .map(s => s.trim());
+      .map(s => s.trim())
   }
 }
 
@@ -162,9 +162,9 @@ async function shWithJson(command, options) {
   const {stdout} = await sh(command, {
     ...options,
     spawnOptions: {stdio: 'pipe', ...(options || {}).spawnOptions},
-  });
+  })
 
-  return JSON.parse(stdout);
+  return JSON.parse(stdout)
 }
 
 module.exports = {
@@ -175,4 +175,4 @@ module.exports = {
   sh,
   shWithOutput,
   shWithJson,
-};
+}