+++ packages/eyes-webdriverio-5-service/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 1.6.4 - 2020/6/29
+
+- updated to @applitools/eyes-webdriverio@5.15.1 (from 5.14.0)
+
 ## 1.6.3 - 2020/6/16
 
 - updated to @applitools/eyes-webdriverio@5.14.0 (from 5.13.2)