+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 5.15.1 - 2020/6/29
+
+- updated to @applitools/visual-grid-client@14.5.0 (from 14.4.10)
+
+## 5.15.0 - 2020/6/29
+
+- remove IosScreenOrientation ([Trello](https://trello.com/c/abSJ68Rl/409-ufg-safari-on-ios-orientations-changes))
+- updated to @applitools/eyes-sdk-core@11.1.0 (from 11.0.10)
+- updated to @applitools/visual-grid-client@14.4.10 (from 14.4.9)
+
 ## 5.14.2 - 2020/6/18
 
 - fix readme